[Chore] #196 - 2차 구현물 파일명 변경
